updated destory of previously used push topics

+++ app/services/salesforce_subscriber.rb
@@ -18,15 +18,15 @@ class SalesforceSubscriber
   def create_account_push_topic
     # Create a PushTopic for subscribing to Account changes.
     delete_account_topics
-    account_topic = @client.create!('PushTopic',
-                                    ApiVersion: '48.0',
-                                    Name: ENV['ACCOUNT_PUSH_TOPIC_NAME'],
-                                    Description: 'all account records',
-                                    NotifyForOperations: 'All',
-                                    NotifyForFields: 'All',
-                                    Query: 'select Id, Name, Phone, Type, Ownerid, ParentId, IsDeleted from Account')
+    account_topic = @client.create('PushTopic',
+                                   ApiVersion: '48.0',
+                                   Name: ENV['ACCOUNT_PUSH_TOPIC_NAME'],
+                                   Description: 'all account records',
+                                   NotifyForOperations: 'All',
+                                   NotifyForFields: 'All',
+                                   Query: 'select Id, Name, Phone, Type, Ownerid, ParentId, IsDeleted from Account')
     
-    PushTopic.create(sf_id: account_topic, topic_type: 'account') if account_topic.present? && account_topic.is_a?(String)
+    PushTopic.create(sf_id: account_topic, is_account: true) if account_topic.present? && account_topic.is_a?(String)
     Logger.new(LOG_PATH).info('Account Push Topic Id: ' + account_topic)
   rescue Restforce::ErrorCode::DuplicateValue
     delete_account_topics
@@ -36,15 +36,15 @@ class SalesforceSubscriber
   def create_contact_push_topic
     # Create a PushTopic for subscribing to Contact changes.
     delete_contact_topics
-    contact_topic = @client.create!('PushTopic',
-                                    ApiVersion: '48.0',
-                                    Name: ENV['CONTACT_PUSH_TOPIC_NAME'],
-                                    Description: 'all contact records',
-                                    NotifyForOperations: 'All',
-                                    NotifyForFields: 'All',
-                                    Query: 'select Id, FirstName, LastName, Title, Email, Phone, MobilePhone, OtherPhone, AccountId, OwnerId, IsDeleted from Contact')
+    contact_topic = @client.create('PushTopic',
+                                   ApiVersion: '48.0',
+                                   Name: ENV['CONTACT_PUSH_TOPIC_NAME'],
+                                   Description: 'all contact records',
+                                   NotifyForOperations: 'All',
+                                   NotifyForFields: 'All',
+                                   Query: 'select Id, FirstName, LastName, Title, Email, Phone, MobilePhone, OtherPhone, AccountId, OwnerId, IsDeleted from Contact')
     
-    PushTopic.create(sf_id: contact_topic, topic_type: 'contact') if contact_topic.present? && contact_topic.is_a?(String)
+    PushTopic.create(sf_id: contact_topic, is_account: false) if contact_topic.present? && contact_topic.is_a?(String)
     Logger.new(LOG_PATH).info('Contact Push Topic Id: ' + contact_topic)
   rescue Restforce::ErrorCode::DuplicateValue
     delete_contact_topics
@@ -70,18 +70,24 @@ class SalesforceSubscriber
   end
   
   def delete_account_topics
-    PushTopic.account.each do |topic|
-      @client.destroy('PushTopic', topic.sf_id)
-      Logger.new(LOG_PATH).info('Account PushTopic destroyed: ' + topic.sf_id)
-      topic.destroy
+    topics = PushTopic.where(is_account: true)
+    if topics.present?
+      topics.each do |topic|
+        @client.destroy('PushTopic', topic.sf_id)
+        Logger.new(LOG_PATH).info('Account PushTopic destroyed: ' + topic.sf_id)
+        topic.destroy
+      end
     end
   end
   
   def delete_contact_topics
-    PushTopic.contact.each do |topic|
-      @client.destroy('PushTopic', topic.sf_id)
-      Logger.new(LOG_PATH).info('Contact PushTopic destroyed: ' + topic.sf_id)
-      topic.destroy
+    topics = PushTopic.where(is_account: false)
+    if topics.present?
+      topics.each do |topic|
+        @client.destroy('PushTopic', topic.sf_id)
+        Logger.new(LOG_PATH).info('Contact PushTopic destroyed: ' + topic.sf_id)
+        topic.destroy
+      end
     end
   end
 end
